net_stubs: Add v_buf_offset argument.
Cstruct buffers may be slices of larger buffers, in which case it is not
sufficient to just pass through the buffer and length; this adds an
additional argument to mirage_solo5_net_read() and
mirage_solo5_net_write(), and computes the correct offset to pass down
to Solo5.

Additionally, to prevent accidental linking with old versions of
mirage-net-solo5, the suffix "_2" has been added to the function
names.