Update to new Solo5 "manifest"-style APIs
This is a minimal update of the C stubs and main loop to use the new
Solo5 "manifest"-style APIs with support for multiple devices.

Note that this change does NOT modify the main loop to support multiple
handles which are ready for input and as such will NOT function
correctly in the presence of multiple network devices in the unikernel.
The functionality for full multiple device support will come in a
subsequent commit.