lib/bindings: Use Store_field() rather than assigning to Field()
While treating Field() as an lvalue and assigning directly to it is
possible, it appears that this is not always safe.

Tests have shown that when storing Int64 values into structured blocks
(tuples) by direct assignment to Field(), the resulting code can fault
under load attempting to access an Int64 at an invalid address (0x1).

Therefore, we will play it safe and use Store_field() throughout.

/cc @stedolan