OS.Memory: new memory statistics interface
Provide a new OS.Memory statistics interface modeled on the stdlib's Gc stat
type. All units are in the system word size.

The implementation takes care to account for the fact that on Solo5 backends the
heap is shared with the stack, and a malloc() allocation on the heap will fail
if the heap would grow to within 1MB of the stack.

Mark OS.MM deprecated

From mirage-xen 38b5a0f831250c26e22f136a96474ee32d6c43c5

Co-authored-by: Martin Lucina <martin@lucina.net>