mirage-solo5 is compatible with opam & MirageOS 4.0
This patch ensures that we are able to `opam install mirage-solo5` and the
right `oS.cmxa` with the right archive `libmirage-solo5_bindings.a` which can
compile only into a specific `dune`'s context.

The idea is to fake an archive when the `ocaml-freestanding`'s toolchain is not
available. In the context of `opam`, we will generate a simple `oS.cmxa` which
expects some undefined symbols.

In the context of `dune` + `opam monorepo`, `dune` should be smart enough to
try to link `libmirage-solo5_bindings.freestanding.a` (which can be compiled
with `ocamlfind -toolchain freestanding opt`) with `oS.cmxa` - and then,
provide missing symbols required to make an unikernel.

NOTE: this patch **does not support** MirageOS 3 - at least, we did not
test this compilation path.